pyproject.toml
[project]
name = "fastapi-template-microservices"
version = "0.1.0"
description = "Template to create simple microservices powered by FastAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3,<2.0.0",
    "asyncpg>=0.31.0,<0.32.0",
    "fastapi[standard]>=0.128.4,<0.129.0",
    "psycopg[binary]>=3.3.2,<4.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "sqlalchemy>=2.0.46,<3.0.0",
    "uvicorn>=0.40.0,<0.41.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "pyjwt>=2.11.0,<3.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "tenacity>=9.1.4,<10.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1,<2.0.0",
    "pytest>=8.3.0,<9.0.0",
    "ruff>=0.15.0,<0.16.0",
    "sqlalchemy[mypy]>=2.0.46,<3.0.0",
    "types-requests>=2.32.4.20260107,<3.0.0.0",
    "pytest-asyncio>=0.25.3,<0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
